Move to a dedicated version module
